+++ packages/benchmarks/src/textToDriver/notebooks-and-scripts/data_out/collections_metadata.json
@@ -1,0 +1,1 @@
+{"dataset": "cinema_recommended", "databaseName": "cinema_recommended", "collectionName": "film", "indexes": [{"v": {"$numberInt": "2"}, "key": {"_id": {"$numberInt": "1"}}, "name": "_id_"}, {"v": {"$numberInt": "2"}, "key": {"Film_ID": {"$numberInt": "1"}, "Directed_by": {"$numberInt": "1"}, "Production_code": {"$numberInt": "1"}, "Title": {"$numberInt": "1"}, "Original_air_date": {"$numberInt": "1"}, "Number_in_season": {"$numberInt": "1"}, "Rank_in_series": {"$numberInt": "1"}}, "name": "Film_ID_1_Directed_by_1_Production_code_1_Title_1_Original_air_date_1_Number_in_season_1_Rank_in_series_1"}, {"v": {"$numberInt": "2"}, "key": {"Film_ID": {"$numberInt": "1"}}, "name": "Film_ID_1"}]}[{"collections": [{"collectionName": "cinema", "indexes": [{"v": {"$numberInt": "2"}, "key": {"_id": {"$numberInt": "1"}}, "name": "_id_"}, {"v": {"$numberInt": "2"}, "key": {"Cinema_ID": {"$numberInt": "1"}, "Capacity": {"$numberInt": "1"}, "Openning_year": {"$numberInt": "1"}, "Name": {"$numberInt": "1"}, "Location": {"$numberInt": "1"}}, "name": "Cinema_ID_1_Capacity_1_Openning_year_1_Name_1_Location_1"}, {"v": {"$numberInt": "2"}, "key": {"Cinema_ID": {"$numberInt": "1"}}, "name": "Cinema_ID_1"}]}, {"collectionName": "film", "indexes": [{"v": {"$numberInt": "2"}, "key": {"_id": {"$numberInt": "1"}}, "name": "_id_"}, {"v": {"$numberInt": "2"}, "key": {"Film_ID": {"$numberInt": "1"}, "Directed_by": {"$numberInt": "1"}, "Production_code": {"$numberInt": "1"}, "Title": {"$numberInt": "1"}, "Original_air_date": {"$numberInt": "1"}, "Number_in_season": {"$numberInt": "1"}, "Rank_in_series": {"$numberInt": "1"}}, "name": "Film_ID_1_Directed_by_1_Production_code_1_Title_1_Original_air_date_1_Number_in_season_1_Rank_in_series_1"}, {"v": {"$numberInt": "2"}, "key": {"Film_ID": {"$numberInt": "1"}}, "name": "Film_ID_1"}]}], "dataset": "cinema_recommended", "databaseName": "cinema_recommended"}]
