+++ packages/ingest-mongodb-public/src/PublicIngestEnvVars.ts
@@ -17,8 +17,18 @@ export const PUBLIC_INGEST_MONGODB_DOCS_META_ENV_VARS = {
   MONGODB_META_DATABASE_NAME: "",
 };
 
+export const INGEST_DEVCENTER_ENV_VARS = {
+  DEVCENTER_CONNECTION_URI: "",
+};
+
+export const INGEST_MDBU_ENV_VARS = {
+  UNIVERSITY_DATA_API_KEY: "",
+};
+
 export const PUBLIC_INGEST_ENV_VARS = {
   ...CORE_CHATBOT_APP_ENV_VARS,
   ...CORE_OPENAI_EMBEDDING_ENV_VARS,
   ...INGEST_ENV_VARS,
+  ...INGEST_DEVCENTER_ENV_VARS,
+  ...INGEST_MDBU_ENV_VARS,
 };
