+++ packages/ingest-mongodb-public/src/sources/snooty/SnootyProjectsInfo.ts
