+++ packages/ingest-mongodb-public/src/sources/snooty/renderSnootyTable.ts
